Pipe-able version of the cat()

The pipe_cat() function allows messages to be output to the console or to an external file without the need to break out of a pipeline to do so.

pipe_cat(
  data,
  ...,
  file = "",
  sep = " ",
  fill = FALSE,
  labels = NULL,
  append = FALSE
)

Arguments

data
the data being passed through the pipeline

...
arguments to be passed to the cat() function. Arguments will be evaluated in the context of data before being passed to cat()

file, sep, fill, labels, append
See the cat() documentation for more information regarding these arguments.

Examples

sample(100,1) %>%
  runif() %>%
  pipe_cat("Current length: ",length(.),"\n") %>%
  pipe_cat("Current average: ",mean(.),"\n") %>%
  pipe_cat("Current standard error: ",sd(.)/length(.),"\n") %>%
  pipe_cat("Returning mean:\n") %>%
  mean

#> Current length:  87 
#> Current average:  0.4873048 
#> Current standard error:  0.003279136 
#> Returning mean:

#> [1] 0.4873048

tibble::tibble(x = runif(10),
               y = runif(10)) %>%
  pipe_cat("Average x: ", mean(x),"\n") %>%
  pipe_cat("Current number of rows: ",nrow,"\n") %>%
  dplyr::mutate(z = x + y)

#> Average x:  0.4680811 
#> Current number of rows:  10 

#> # A tibble: 10 x 3
#>        x      y     z
#>    <dbl>  <dbl> <dbl>
#>  1 0.927 0.928  1.86 
#>  2 0.894 0.0266 0.921
#>  3 0.204 0.559  0.763
#>  4 0.257 0.854  1.11 
#>  5 0.614 0.0759 0.690
#>  6 0.441 0.206  0.647
#>  7 0.316 0.496  0.811
#>  8 0.101 0.262  0.363
#>  9 0.273 0.416  0.689
#> 10 0.654 0.439  1.09 

iris %>%
  dplyr::mutate(Species = as.character(Species)) %>%
  pipe_cat("Total average Sepal Length: ",mean(Sepal.Length),"\n") %>%
  dplyr::group_by(Species) %>%
  pipe_cat(Species," average Sepal Length: ",mean(Sepal.Length),"\n")

#> Total average Sepal Length:  5.843333 

#> Warning: The `keep` argument of `group_split()` is deprecated as of dplyr 1.0.0.
#> Please use the `.keep` argument instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> setosa  average Sepal Length:  5.006 
#> versicolor  average Sepal Length:  5.936 
#> virginica  average Sepal Length:  6.588 

#> # A tibble: 150 x 5
#> # Groups:   Species [3]
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>           <dbl>       <dbl>        <dbl>       <dbl> <chr>  
#>  1          5.1         3.5          1.4         0.2 setosa 
#>  2          4.9         3            1.4         0.2 setosa 
#>  3          4.7         3.2          1.3         0.2 setosa 
#>  4          4.6         3.1          1.5         0.2 setosa 
#>  5          5           3.6          1.4         0.2 setosa 
#>  6          5.4         3.9          1.7         0.4 setosa 
#>  7          4.6         3.4          1.4         0.3 setosa 
#>  8          5           3.4          1.5         0.2 setosa 
#>  9          4.4         2.9          1.4         0.2 setosa 
#> 10          4.9         3.1          1.5         0.1 setosa 
#> # … with 140 more rows